+++ watchtower/__init__.py
@@ -2,5 +2,5 @@
 WatchTower - Podman Container Management Service
 """
 
-__version__ = "1.16.0"
+__version__ = "1.16.1"
 __author__ = "The WatchTower Authors"